+++ test-projects/access-control/tests/queries/meta.test.js
@@ -1,5 +1,5 @@
-const { Text, Relationship } = require('@voussoir/fields');
-const { multiAdapterRunners, setupServer, graphqlRequest } = require('@voussoir/test-utils');
+const { Text, Relationship } = require('@keystone-alpha/fields');
+const { multiAdapterRunners, setupServer, graphqlRequest } = require('@keystone-alpha/test-utils');
 const cuid = require('cuid');
 
 function setupKeystone(adapterName) {