+++ packages/fields/types/Content/views/editor/index.js
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import { useRef, useMemo } from 'react';
+import { useMemo } from 'react';
 import { Editor } from 'slate-react';
 import { Block } from 'slate';
 import { plugins as markPlugins } from './marks';
@@ -67,9 +67,8 @@ function Stories({ value: editorState, onChange, blocks, className }) {
   );
 
   let [editor, setEditor] = useStateWithEqualityCheck(null);
-  let containerRef = useRef(null);
   return (
-    <div ref={containerRef} className={className}>
+    <div className={className}>
       <Editor
         schema={schema}
         ref={setEditor}